requirements.txt
fastapi
uvicorn[standard]
orjson
sqlalchemy>=2.0
psycopg2-binary
alembic
pydantic>=2
pydantic-settings
python-dotenv
openai>=1.40.0
rapidfuzz
pgvector
prometheus-client
sentry-sdk
reportlab
boto3
python-multipart
xlsxwriter
